refactor after conversation
