+++ frontend/src/pages/SkillsMap/SkillsMap.tsx
@@ -7,18 +7,16 @@ import { ISkillsMapProps } from 'pages/SkillsMap/types';
 import StackMap from './StackMap';
 import { SkillsWrapper } from './styled';
 
-const SkillsMap: FC<ISkillsMapProps> = ({ coursesMapResponce, isCoursesMapResponceLoading }) => {
-  return (
-    <AuthorizedLayout pageName="Skills Map">
-      <SkillsWrapper isLoading={isCoursesMapResponceLoading}>
-        {isCoursesMapResponceLoading ? (
-          <Loader color="primary" type="component" />
-        ) : (
-          <StackMap coursesMapResponce={coursesMapResponce} />
-        )}
-      </SkillsWrapper>
-    </AuthorizedLayout>
-  );
-};
+const SkillsMap: FC<ISkillsMapProps> = ({ coursesMapResponse, isCoursesMapLoading }) => (
+  <AuthorizedLayout pageName="Skills Map">
+    <SkillsWrapper isLoading={isCoursesMapLoading}>
+      {isCoursesMapLoading ? (
+        <Loader color="primary" type="component" />
+      ) : (
+        <StackMap coursesMapResponce={coursesMapResponse} />
+      )}
+    </SkillsWrapper>
+  </AuthorizedLayout>
+);
 
 export default SkillsMap;
