--- frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/Courses/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Courses';
