+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/CoursesList/Course/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CourseContainer';
