+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/CoursesList/Course/styled.ts
@@ -33,7 +33,3 @@ export const CourseTitle = styled('span')({
   textOverflow: 'ellipsis',
   overflow: 'hidden',
 });
-
-export const NoCoursesText = styled('h4')({
-  margin: '10px 0',
-});
