+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/CoursesList/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CoursesList';
