+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/CoursesList/styled.ts
@@ -1,0 +1,5 @@
+import { styled } from '@mui/material';
+
+export const NoCoursesText = styled('h4')({
+  margin: '10px 0',
+});
