+++ frontend/src/pages/SkillsMap/StackMap/CourseMapForCurrentStack/styled.ts
@@ -9,15 +9,12 @@ export const Rank = styled('div')<{ selected?: boolean }>(({ selected }) => ({
   padding: '20px',
   borderRadius: '10px',
   backgroundColor: '#F1F1F1',
-
   ...(selected && {
     border: '4px solid rgb(212 62 65 / 20%)',
   }),
-
   '&:not(:last-child)': {
     marginRight: '40px',
   },
-
   [theme.breakpoints.down('lg')]: {
     width: '100%',
     margin: '10px 0',
@@ -35,11 +32,9 @@ export const CoursesWrapper = styled('div')({
   flexDirection: 'column',
   maxHeight: '440px',
   overflowY: 'scroll',
-
   [theme.breakpoints.down('xl')]: {
     maxHeight: '365px',
   },
-
   [theme.breakpoints.down('lg')]: {
     maxHeight: '100%',
   },
