+++ frontend/src/types/course.ts
@@ -1,5 +1,6 @@
 import { COURSE_LABELS } from 'constants/statuses';
 import { CourseStatus } from 'enums/course';
+import { EmployeeRank } from 'enums/employee';
 
 import { TimeProps } from './time';
 
@@ -77,4 +78,27 @@ export interface ICourseInfo {
   duration?: string;
   lessons?: number;
   type?: string;
+}
+
+export interface ICoursesMapResponse {
+  userRank: EmployeeRank;
+  stackMap: IStackMapElement[];
+}
+
+export interface IStackMapElement {
+  coursesMap: ICoursesMapElement[];
+  stack: string;
+  isPrimary: boolean;
+}
+
+export interface ICoursesMapElement {
+  courses: ICourseShortInfo[];
+  rank: EmployeeRank;
+}
+
+export interface ICourseShortInfo {
+  _id: string;
+  avatar: string;
+  title: string;
+  isCompleted: boolean;
 }
